Star citzen server/requirements_google_apis.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
googleapis-common-protos>=1.56.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-api-core>=2.11.0
asyncio